# Backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: FixFlow Backend API (Aggregate)
  version: '1.0'
  description: |
    OpenAPI specification (gateway-level) for Users and Orders APIs (v1).
    Security is handled at the gateway: all protected endpoints require a Bearer JWT.
servers:
  - url: http://localhost:8000/v1
tags:
  - name: auth
    description: Authentication endpoints (registration, login)
  - name: users
    description: User profile and management
  - name: orders
    description: Orders lifecycle
paths:
  /auth/register:
    post:
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
            examples:
              basic:
                value: {"email": "alice@example.com", "password": "s3cret", "name": "Alice"}
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPublic'
        '400':
          $ref: '#/components/responses/BadRequest'
  /auth/login:
    post:
      tags: [auth]
      summary: Obtain an access token (JWT)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
            examples:
              basic:
                value: {"email": "alice@example.com", "password": "s3cret"}
      responses:
        '200':
          description: Access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/me:
    get:
      tags: [users]
      summary: Get current user's profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPublic'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users:
    get:
      tags: [users]
      summary: List users (requires admin role)
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Paginated list of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserPublic'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /orders:
    post:
      tags: [orders]
      summary: Create a new order
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreate'
            examples:
              sample:
                value: {"items": [{"sku":"ABC123","qty":2,"price":12.5}], "total": 25.0}
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /orders/{order_id}:
    get:
      tags: [orders]
      summary: Get order by id
      parameters:
        - in: path
          name: order_id
          required: true
          schema:
            type: string
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Order found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Unauthorized:
      description: Authentication required or invalid credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Authenticated but not allowed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        detail:
          type: string
      required: [detail]
    UserCreate:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        name:
          type: string
      required: [email, password]
    UserLogin:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    UserPublic:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
        roles:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
      required: [id, email]
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
          example: bearer
        expires_in:
          type: integer
          description: seconds until expiration
      required: [access_token, token_type]
    OrderCreate:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
            properties:
              sku:
                type: string
              qty:
                type: integer
              price:
                type: number
            required: [sku, qty, price]
        total:
          type: number
      required: [items, total]
    Order:
      type: object
      properties:
        id:
          type: string
        user_id:
          type: string
        items:
          type: array
          items:
            type: object
        total:
          type: number
        status:
          type: string
          description: Order status (e.g., created, processing, completed, cancelled)
        created_at:
          type: string
          format: date-time
      required: [id, user_id, items, total, status]